+++ micro-focus-service-manager/release_notes.md
@@ -1,5 +1,8 @@
 #### What's Added
 - Added the following new actions and its playbooks
-  - Create Change
-  - Get Change List
-  - Get Change Request
+  
+  - Update Change
+  - Get RF - Request Fulfillment Ticket
+  - Create RF - Request Fulfillment Ticket
+  - Update RF - Request Fulfillment Ticket for an attachment
+